Struct chunky::Value

pub struct Value<V, H: Handler> { /* fields omitted */ }

A single value stored in a chunk

Methods

impl<V, H: Handler> Value<V, H>

pub fn load_or_default(ident: Ident, default: V) -> Value<V, H>

Load the value in the chunk with the given identifier, or create it using a default value